+++ packages/rx-display/src/lib/badge/badge.tsx
@@ -31,6 +31,8 @@ const badgeVariants = cva(
           'bg-secondary text-secondary-foreground [a&]:hover:bg-secondary/90',
         destructive:
           'bg-destructive text-white focus-visible:ring-destructive/20 dark:bg-destructive/60 dark:focus-visible:ring-destructive/40 [a&]:hover:bg-destructive/90',
+        warning: 'bg-warning text-warning-foreground [a&]:hover:bg-warning/90',
+        success: 'bg-success text-success-foreground [a&]:hover:bg-success/90',
         outline:
           'border-border text-foreground [a&]:hover:bg-accent [a&]:hover:text-accent-foreground',
         ghost: '[a&]:hover:bg-accent [a&]:hover:text-accent-foreground',