+++ packages/gremorie/CHANGELOG.md
@@ -1,5 +1,11 @@
 # gremorie
 
+## 0.7.0
+
+### Patch Changes
+
+- @gremorie/cli@0.7.0
+
 ## 0.6.0
 
 ### Patch Changes