+++ packages/ng-artifacts/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @gremorie/ng-artifacts
 
+## 0.7.0
+
 ## 0.6.0
 
 ## 0.5.1