+++ packages/ng-core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @gremorie/ng-core
 
+## 0.7.0
+
 ## 0.6.0
 
 ### Patch Changes