+++ packages/ng-navigation/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @gremorie/ng-navigation
 
+## 0.7.0
+
 ## 0.6.0
 
 ### Patch Changes