+++ packages/ng-overlays/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @gremorie/ng-overlays
 
+## 0.7.0
+
 ## 0.6.0
 
 ### Patch Changes