+++ packages/rx-overlays/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @gremorie/rx-overlays
 
+## 0.7.0
+
 ## 0.6.0
 
 ## 0.5.1